requirements.txt
alembic==1.4.2
cffi==1.14.0
click==7.1.1
cryptography==2.8
docutils==0.15.2
Flask==1.1.1
Flask-Cors==3.0.8
flask-marshmallow==0.11.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
gunicorn==20.0.4
itsdangerous==1.1.0
PyMySQL==0.9.3
Jinja2==2.11.1
jmespath==0.9.5
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.3
mysqlclient==1.4.6
Pillow==7.1.1
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
s3transfer==0.3.3
six==1.14.0
SQLAlchemy==1.3.15
SQLAlchemy-Utils==0.36.3
urllib3==1.25.9
Werkzeug==1.0.0
